pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = "test Backend"
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.7"
django-model-utils = "^4.3.1"
django-ninja = "^0.22.2"
django-cities-light = "^3.10.0"
django-money = "^3.4.0"
django-auditlog = "^2.3.0"
pydantic = "^1.10"
email-validator = "^2.1.0.post1"
pyyaml = "^6.0.1"
django-guardian = "^2.4.0"
django-phonenumber-field = "^7.2.0"
phonenumbers = "^8.13.26"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-django = "^4.7.0"
cookiecutter = "^2.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
